+++ src/pages/NwbPage/plugins/registry.ts
@@ -14,9 +14,11 @@ import { trialAlignedSeriesPlugin } from "./TrialAlignedSeries";
 import { pythonScriptPlugin } from "./PythonScript";
 import spikeDensityPlugin from "./SpikeDensity";
 import { intervalSeriesPlugin } from "./IntervalSeries";
+import { eventsPlugin } from "./Events";
 
 // List of plugins in order they will appear in the UI when a single object is being viewed
 export const nwbObjectViewPlugins: NwbObjectViewPlugin[] = [
+  eventsPlugin,
   intervalSeriesPlugin,
 
   behavioralEventsPlugin,